+++ src/H5FSstat.c
@@ -96,7 +96,7 @@ H5FS_stat_info(const H5F_t *f, const H5FS_t *frsp, H5FS_stat_t *stats)
     stats->serial_sect_count = frsp->serial_sect_count;
     stats->ghost_sect_count = frsp->ghost_sect_count;
     stats->addr = frsp->addr;
-    stats->hdr_size = H5FS_HEADER_SIZE(f);
+    stats->hdr_size = (size_t)H5FS_HEADER_SIZE(f);
     stats->sect_addr = frsp->sect_addr;
     stats->alloc_sect_size = frsp->alloc_sect_size;
     stats->sect_size = frsp->sect_size;
